adding new cv route and installed next intl
