+++ client/messages/fr.json
@@ -1,0 +1,4 @@
+{
+
+    
+}
