+++ client/next.config.mjs
@@ -1,4 +1,10 @@
 /** @type {import('next').NextConfig} */
+
+
+import createNextIntlPlugin from 'next-intl/plugin';
+
+const withNextIntl = createNextIntlPlugin();
+
 const nextConfig = {
 
    
@@ -29,4 +35,4 @@ const nextConfig = {
 
 };
 
-module.exports = nextConfig
+export default withNextIntl(nextConfig)
