+++ client/package.json
@@ -9,7 +9,10 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "framer-motion": "^11.11.7",
+    "monorepo": "github:radix-ui/themes",
     "next": "13.4.19",
+    "next-intl": "^3.21.1",
     "react": "18.2.0",
     "react-dom": "18.2.0"
   }
