+++ client/src/app/[locale]/_components/bio.js
