+++ client/src/app/[locale]/_components/single-projects/singleProject.js
