+++ client/src/app/[locale]/_components/single-projects/singleProject.module.css
