+++ client/src/app/[locale]/apple-icon.png
