+++ client/src/app/[locale]/cv/page.jsx
@@ -1,0 +1,12 @@
+
+
+
+
+export default function Page(){
+
+    return(
+        <>
+            <p>this</p>
+        </>
+    )
+}
