+++ client/src/app/[locale]/favicon.png
