+++ client/src/app/[locale]/globals.css
@@ -11,22 +11,9 @@ body {
   max-width: 100vw;
   overflow-x: hidden;
   
-  color: rgb(255, 255, 255);
-}
-
-body {
-  
-  /*background-image: url('../public/images/cascade.jpg');
-  background-repeat: no-repeat;
-  background-size: cover ;*/
-  background-image: linear-gradient(to right, #ffffff, #907ff3);
-  color: rgb(255, 255, 255);
-  display: flex;
-  flex-direction: column;
-  margin: 0px;
-  
 
 }
+
 
 a {
   color: inherit;
@@ -38,12 +25,12 @@ h1{
 }
 
 h2{
-  font-size: 1.9em;
+  font-size: 1.6em;
   font-weight: 500;
 }
 
 p {
-  font-size: 1.5em;
+  font-size: 1.2em;
   font-weight: 400;
 }
 
@@ -68,7 +55,8 @@ p {
 
 footer{
   display: flex;
-  justify-content: space-around;
+  width: 90vw;
+  justify-content: space-between;
   padding: 5%;
   font-weight: 600;
   color: #59595c;
