+++ client/src/app/[locale]/icon.png
