+++ client/src/i18n.js
@@ -1,0 +1,13 @@
+import { notFound } from "next/navigation";
+import { getRequestConfig } from 'next-intl/server';
+
+const locales = ['en', 'fr'];
+
+export default getRequestConfig(async ({ locale }) => {
+  if (!locales.includes(locale)) notFound();
+
+  return {
+    messages: (await import(`../messages/${locale}.json`)).default
+  };
+});
+
